edit frontend add authorization system, home, single post, new post, update post, fix some api in golang, edit sql model structures

+++ server/controller/commentController.go
@@ -4,10 +4,12 @@ import (
 	"errors"
 	"fmt"
 	"strconv"
+	"time"
 
 	"github.com/gofiber/fiber/v2"
 	"github.com/purivirakarin/cvwo-assignment-2023/server/database"
 	"github.com/purivirakarin/cvwo-assignment-2023/server/models"
+	"github.com/purivirakarin/cvwo-assignment-2023/server/util"
 	"gorm.io/gorm"
 )
 
@@ -16,53 +18,78 @@ func CreateComment(c *fiber.Ctx) error {
 	if err := c.BodyParser(&comment); err != nil {
 		fmt.Println("Unable to parse body")
 	}
+	comment.Date = time.Now().Unix()
+	cookie := c.Cookies("jwt")
+	id, _ := util.Parsejwt(cookie)
+	comment.UserId = id
 	if err := database.DB.Create(&comment).Error; err != nil {
 		c.Status(400)
 		return c.JSON(fiber.Map{
-			"message": "Invalid payload",
+			"message": "Invalid payload.",
 		})
 	}
 	return c.JSON(fiber.Map{
-		"message": "Congratulation!, You post is live!",
+		"message": "Successfully created a comment.",
 	})
 }
 
 func UpdateComment(c *fiber.Ctx) error {
+	cookie := c.Cookies("jwt")
+	userId, _ := util.Parsejwt(cookie)
+
 	id, _ := strconv.Atoi(c.Params("id"))
-	comment := models.Comment{
-		Id: uint(id),
+	var comment models.Comment
+	database.DB.Model(&comment).Where("id=?", id).First(&comment)
+	if comment.UserId != userId {
+		c.Status(401)
+		return c.JSON(fiber.Map{
+			"message": "Unauthorized access.",
+		})
 	}
 	if err := c.BodyParser(&comment); err != nil {
 		fmt.Println("Unable to parse body")
 	}
 	database.DB.Model(&comment).Updates(comment)
-	return c.JSON(fiber.Map{
-		"message": "Post updated successfully",
-		"data":    comment,
-	})
+	return c.JSON(comment)
 }
 
 func UniqueComment(c *fiber.Ctx) error {
 	id, _ := strconv.Atoi(c.Params("id"))
 	var comment []models.Comment
-	database.DB.Model(&comment).Where("forum_id=?", id).Preload("Forum").Find(&comment)
+	database.DB.Model(&comment).Where("forum_id=?", id).Preload("User").Preload("Forum").Find(&comment)
 	return c.JSON(comment)
 }
 
 func DeleteComment(c *fiber.Ctx) error {
+	cookie := c.Cookies("jwt")
+	userId, _ := util.Parsejwt(cookie)
+
 	id, _ := strconv.Atoi(c.Params("id"))
-	comment := models.Comment{
-		Id: uint(id),
+	var comment models.Comment
+	database.DB.Model(&comment).Where("id=?", id).First(&comment)
+	if comment.UserId != userId {
+		c.Status(401)
+		return c.JSON(fiber.Map{
+			"message": "Unauthorized access.",
+		})
 	}
 	deleteQuery := database.DB.Delete(&comment)
 	if errors.Is(deleteQuery.Error, gorm.ErrRecordNotFound) {
-		c.Status(400)
+		c.Status(500)
 		return c.JSON(fiber.Map{
 			"message": "Opps!, the post is invalid",
 		})
 	}
 
 	return c.JSON(fiber.Map{
-		"message": "post deleted succesfully",
+		"message": "Post deleted succesfully.",
 	})
+}
+
+func CommentByForumId(c *fiber.Ctx) error {
+	forumId, _ := strconv.Atoi(c.Params("forumid"))
+
+	var comments []models.Comment
+	database.DB.Where("forum_id=?", forumId).Preload("User").Preload("Forum").Find(&comments)
+	return c.JSON(comments)
 }
